% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_joint.R
\name{create_empty_joint}
\alias{create_empty_joint}
\title{Create empty joint from targets}
\usage{
create_empty_joint(targets)
}
\arguments{
\item{targets}{}
}
\description{
Create empty joint from targets
}
