% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{do_boot_ma}
\alias{do_boot_ma}
\title{Estimate the model averaged ATE on one bootstrap sample}
\usage{
do_boot_ma(data0, endpoints, categorical, treatment, ...)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
A list with point estimates using BIC model averagin and
  sub-model estimates, SE estimates, and weights,
}
\description{
Estimate the model averaged ATE on one bootstrap sample
}
