% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{estimate_bma}
\alias{estimate_bma}
\title{Estimate the ATE using the BMA-SEM estimator}
\usage{
estimate_bma(
  data0,
  endpoints,
  treatment = "A",
  primary,
  ci_level = 0.95,
  force_probit = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{ci_level}{Confidence level}

\item{force_probit}{For binary endpoints only. If TRUE, reports probit
regression coefficient instead of risk difference.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
A vector of point estimates using the saturated, SEM, and BMA
estimators, variance estimates, and and BMA model weights
}
\description{
\code{estimate_bma} estimates the ATE and conditional means of the primary
endpoint on one dataset using Bayesian model averaging.
}
