% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{estimate_ma}
\alias{estimate_ma}
\title{Estimate the ATE and conditional means using the saturated, SEM, BMA, and SL
estimators}
\usage{
estimate_ma(
  data0,
  endpoints,
  treatment = "A",
  primary,
  n_boot = 100,
  ci_level = 0.95,
  sandwich = FALSE,
  force_probit = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{n_boot}{Number of bootstrap samples}

\item{ci_level}{Confidence level}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimate in addition to the model-based estimator.}

\item{force_probit}{For binary endpoints only. If TRUE, reports probit
regression coefficient instead of risk difference.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
A vector of point estimates using the saturated, SEM, BMA, and SL
estimators, variance estimates, and BMA and SL model weights
}
\description{
Estimate the ATE and conditional means using the saturated, SEM, BMA, and SL
estimators
}
