% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{estimate_ma_2}
\alias{estimate_ma_2}
\title{Estimate the ATE using model averaging}
\usage{
estimate_ma_2(
  data0,
  endpoints,
  categorical,
  treatment = "A",
  sandwich = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimator in addition to the model-based estimator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
A list with point estimates using SL and BIC model averaging,
  sub-model estimates, and weights.
}
\description{
Estimate the ATE using model averaging
}
