% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{estimate_ma_3}
\alias{estimate_ma_3}
\title{Estimate the ATE using model averaging}
\usage{
estimate_ma_3(
  data0,
  endpoints,
  categorical = c(),
  treatment = "A",
  primary,
  n_boot = 100,
  ci_level = 0.95,
  sandwich = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{n_boot}{Number of boostrap samples}

\item{ci_level}{Confidence level}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimator in addition to the model-based estimator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
A list with point estimates using BIC model averaging and
  sub-model estimates, SE estimates, and weights.
}
\description{
Estimate the ATE using model averaging
}
