% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sem.R
\name{estimate_means}
\alias{estimate_means}
\title{Estimate conditional means  from a joint model}
\usage{
estimate_means(model, sandwich = FALSE)
}
\arguments{
\item{model}{An object of class [joint_sem]}

\item{sandwich}{Logical indicator to use the sandwich variance estimator}
}
\value{
A list of mean estimates and variances
}
\description{
Estimate conditional means  from a joint model
}
\examples{
data(joint_example)
fit <- joint_sem(
  joint_example, endpoints = c("Y1", "Y2", "Y3_cat"))
fit
estimate_effects(fit, risk_difference = "Y3_cat")
}
