% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{estimate_sem_tmle}
\alias{estimate_sem_tmle}
\title{Estimate an average treatment effect using TMLE (sl3)}
\usage{
estimate_sem_tmle(
  data0,
  endpoints,
  categorical = c(),
  treatment = "A",
  primary,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{...}{Additional arguments to \code{joint_sem}}
}
\description{
Estimate an average treatment effect using TMLE (sl3)
}
\examples{
data(joint_example)
estimate_ma <- estimate_sem_tmle(
  data0 = joint_example,
  endpoints = c("Y1", "Y2", "Y3"),
  categorical = c(),
  treatment = "A",
  primary = "Y1"
)
estimate_ma$tmle_est
estimate_ma$tmle_sigma
estimate_sem <- estimate_effects(estimate_ma$joint_sem)
estimate_sem$estimate["Y1"]
sqrt(diag(estimate_sem$vcov)["Y1"])
}
