% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit}
\usage{
fit(joint, targets, iterations = 10, remove_zeroes = TRUE)
}
\arguments{
\item{remove_zeroes}{Often cells with zero people in are not required in an analysis.}
}
\description{
Fit
}
