% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marginal.R
\name{get_marginal}
\alias{get_marginal}
\title{Get marginals from a joint distribution}
\usage{
get_marginal(joint, ...)
}
\arguments{
\item{...}{}
}
\description{
Get marginals from a joint distribution
}
