% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{initalize_phi}
\alias{initalize_phi}
\title{Create initial vector of parameter values for estimating a joint model}
\usage{
initalize_phi(data0, endpoints, categorical = c(), gamma = 2)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{gamma}{Initial value of the gamma parameter, defaults to 2}
}
\value{
A named vector of initial parameter values for likelihood
  maximization.
}
\description{
Create initial vector of parameter values for estimating a joint model
}
