% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{joint_bma}
\alias{joint_bma}
\title{Estimate the ATE using the BMA-SEM estimator}
\usage{
joint_bma(
  data0,
  endpoints,
  treatment = "A",
  primary,
  n_boot = 100,
  ci_level = 0.95,
  force_probit = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{n_boot}{Number of bootstrap samples}

\item{ci_level}{Confidence level}

\item{force_probit}{For binary endpoints only. If TRUE, reports probit
regression coefficient instead of risk difference.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
A list containing the following named objects:
\item{summary}{A table with summarized output including ATE estimates,
standard error estimates, and bootstrapped percentile CI bounds}
\item{estimate}{A vector of parameter estimates and diagnostic
information}
\item{boot_results}{A \code{data.frame} where each row is an estimate
obtained from a particular bootstrapped resample of \code{data0}}
\item{runtime}{The run time in seconds}
\item{call}{The matched call}
}
\description{
\code{joint_bma} estimates the ATE and conditional means of the primary
endpoint by Bayesian model averaging between the saturated model and
one-factor structural equation model.
}
\examples{
\donttest{
data(joint_example)
set.seed(1)
joint_bma(
  data0 = joint_example,
  endpoints = c("Y1", "Y2", "Y3"),
  treatment = "A",
  primary = "Y1",
  n_boot = 5,
  ci_level = 0.95
  )
}

}
\references{
{ \insertRef{wolf_jointly_2025}{joint} }
}
\seealso{
\code{\link[=joint_sl]{joint_sl()}}
}
