% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joint_example}
\alias{joint_example}
\title{Example data for fitting joint endpoint models}
\format{
A data frame with 200 observations of 7 variables:
\describe{
 \item{A}{Binary treatment indicator}
 \item{Y1, Y2, Y3, Y4}{Continuous endpoints}
 \item{Y3_cat, Y4_cat}{Categorical versions of Y3 and Y4}
}
}
\usage{
joint_example
}
\description{
Example data for fitting joint endpoint models
}
\keyword{datasets}
