% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_wrappers.R
\name{joint_ma_sim}
\alias{joint_ma_sim}
\title{Estimate the ATE using the saturated, SEM, BMA-SEM, and SL-SEM estimators}
\usage{
joint_ma_sim(
  data0,
  endpoints,
  treatment = "A",
  primary,
  n_boot = 100,
  ci_level = 0.95,
  sandwich = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{primary}{Name of the primary endpoint}

\item{n_boot}{Number of bootstrap replications to perform}

\item{ci_level}{The confidence level required}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimate in addition to the model-based estimator.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
A data.frame with point estimates, standard error estimates, and
bootstrapped percentile confidence interval bounds
}
\description{
Estimate the ATE using the saturated, SEM, BMA-SEM, and SL-SEM estimators
}
\examples{
\donttest{
data(joint_example)
# Using n_boot = 5 for example purposes only, use more bootstrap iterations
# in practice!
# Continuous endpoint (ATE)
joint_ma_sim(
  data0 = joint_example,
  endpoints = c("Y1", "Y2", "Y3_cat"),
  treatment = "A",
  primary = "Y1",
  n_boot = 1)

# Binary endpoint (risk difference ATE)
joint_ma_sim(
  data0 = joint_example,
  endpoints = c("Y1", "Y2", "Y3_cat"),
  treatment = "A",
  primary = "Y3_cat",
  n_boot = 1)
}
}
