% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{joint_ma_sim_2}
\alias{joint_ma_sim_2}
\title{Estimate the ATE with various methods for simulation studies}
\usage{
joint_ma_sim_2(
  data0,
  endpoints,
  categorical = c(),
  treatment = "A",
  primary,
  sandwich = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimator in addition to the model-based estimator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
A data.frame with point estimates for various methods
}
\description{
Estimate the ATE with various methods for simulation studies
}
\examples{
data(joint_example)
joint_ma_sim(joint_example, c("Y1", "Y2", "Y3"), n_boot = 5)
 @examples
  data(joint_example)
  estimate_ma_2(
    joint_example,
    endpoints = c("Y1", "Y2", "Y3"),
    categorical = c(),
    treatment = "A"
   )
}
