% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_saturated.R
\name{joint_saturated}
\alias{joint_saturated}
\title{Estimate a saturated joint model}
\usage{
joint_saturated(data0, endpoints, treatment = "A", ...)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
An object of class \link{joint_sem}

An object of class {"joint_sem"} is a list containing at least the
following components:
\item{estimate}{The vector of estimated model parameters}
\item{ll}{The maximized log likelihood}
\item{dim_phi}{The number of model parameters}
\item{runtime}{The run time in seconds}
\item{endpoints}{The input vector of endpoints}
\item{categorical}{The vector of endpoints treated as categorical}
\item{V_model}{The model-based parameter variance estimate}
\item{nlm_code}{An integer indicating why the optimization process
terminated; see {"code"} from \code{\link[stats:nlm]{stats::nlm()}}}
\item{nlm_iterations}{The number of \code{\link[stats:nlm]{stats::nlm()}} iterations performed}
\item{call}{The matched call}
}
\description{
\code{joint_saturated} fits a saturated joint model for all endpoints. Marginal
distributions are estimated through either linear or probit regression. The
appropriate variance and covariance terms are estimated through maximum
likelihood and the log likelihood is returned.
}
