% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{joint_vglm_sl}
\alias{joint_vglm_sl}
\title{Wrapper function for \code{VGAM::vglm} for super learning.}
\usage{
joint_vglm_sl(x, y, yname, treatment, ...)
}
\arguments{
\item{x}{Matrix of secondary endpoints and treatment indicator}

\item{y}{Vector of primary endpoint measurements}

\item{yname}{Character name for primary endpoint}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}
}
\value{
An object of class \link[VGAM:vglm]{VGAM::vglm}
}
\description{
Wrapper function for \code{VGAM::vglm} for super learning.
}
