% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_saturated.R
\name{ll_cat_saturated}
\alias{ll_cat_saturated}
\title{Calculate the log likelihood for categorical endpoints}
\usage{
ll_cat_saturated(
  phi,
  data0,
  endpoints,
  categorical,
  treatment,
  continuous,
  lb,
  ub,
  mu,
  Sigma
)
}
\arguments{
\item{phi}{Vector of covariance parameter estimates}

\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{lb, ub}{Matrix of lower and upper bounds for integration}

\item{mu}{Matrix of endpoint means (on the probit scale for categorical
endpoints)}

\item{Sigma}{Matrix of endpoint covariances. Some entries are overwritten by
`phi` before likelihood calculation.}
}
\value{
The log likelihood for the input parameters and data
}
\description{
`ll_cat_saturated()` calculates the conditional log likelihood for all
categorical endpoints (conditional on all continuous endpoints) for input
marginal mean and variance matrices. As this function is used internally for
likelihood maximization, the primary input (`phi`) is a vector of covariance
terms for all continuous by categorical and categorical by categorical
covariances.
}
