% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{predict.joint_sem}
\alias{predict.joint_sem}
\title{Predict method for \code{joint_sem} for Super Learning.}
\usage{
\method{predict}{joint_sem}(object, newdata)
}
\arguments{
\item{object}{An object of class \code{joint_sem}}

\item{newdata}{An optional data frame with treatment indicators. If omitted
the treatment indicators from the fitted model are used.}
}
\value{
Predicted outcomes given treatment indicators.
}
\description{
Predict method for \code{joint_sem} for Super Learning.
}
