% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_joint.R
\name{create_empty_joint}
\alias{create_empty_joint}
\title{Create empty joint from targets}
\usage{
create_empty_joint(targets, columns_to_exclude = NULL, backend = "base")
}
\arguments{
\item{backend}{"base", "dplyr" or "datatable"}
}
\description{
Create empty joint from targets
}
