% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit}
\usage{
fit(
  targets,
  empty_joint = NULL,
  iterations = 10,
  remove_zeroes = TRUE,
  backend = "base"
)
}
\arguments{
\item{remove_zeroes}{If cells with zero people are not required for your analysis,
removing zeroes will speed up the fit}
}
\description{
Create a joint distribution from a list of target distributions
}
