% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_model_average.R
\name{joint_ma_sim}
\alias{joint_ma_sim}
\title{Estimate the ATE with various methods for simulation studies}
\usage{
joint_ma_sim(
  data0,
  endpoints,
  categorical = c(),
  treatment = "A",
  nboot = 100,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
A data.frame with point estimates, standard error estimates, test
  statistics, and 95% confidence interval bounds
}
\description{
Estimate the ATE with various methods for simulation studies
}
