% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_saturated.R
\name{joint_saturated}
\alias{joint_saturated}
\title{Estimate a saturated joint model}
\usage{
joint_saturated(data0, endpoints, categorical = c(), treatment = "A", ...)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
An object of class [joint_sem]
}
\description{
`joint_saturated` fits a saturated joint model for all endpoints. Marginal
distributions are estimated through either linear or probit regression. The
appropriate variance and covariance terms are estimated through maximum
likelihood and the log likelihood is returned.
}
\examples{
data(joint_example)
fit <- joint_saturated(
  joint_example, endpoints = c("Y1", "Y2", "Y3_cat"), categorical = "Y3_cat")
fit
}
