% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{joint_sem}
\alias{joint_sem}
\title{Estimate a joint model using SEM constraints}
\usage{
joint_sem(
  data0,
  endpoints,
  categorical = c(),
  sandwich = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimator in addition to the model-based estimator.}

\item{debug}{Logical indicator to print diagnoistic information}

\item{...}{additional arguments to [stats::nlm()].}

\item{endpints}{Character vector of endpoint names as in \code{data0}}
}
\value{
An object of class [joint_sem]
}
\description{
Estimate a joint model using SEM constraints
}
\examples{
data(joint_example)
fit <- joint_sem(
  joint_example, endpoints = c("Y1", "Y2", "Y3_cat"), categorical = "Y3_cat")
fit
estimate_effects(fit, risk_difference = "Y3_cat")
}
