% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sem.R
\name{joint_sem}
\alias{joint_sem}
\title{Estimate a joint model using SEM constraints}
\usage{
joint_sem(
  data0,
  endpoints,
  treatment = "A",
  sandwich = FALSE,
  debug = FALSE,
  phi_init = NULL,
  ...
)
}
\arguments{
\item{data0}{Data frame with treatment assignment and endpoints}

\item{endpoints}{Character vector of endpoint names as in \code{data0}}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{sandwich}{Logical indicator for whether to return the sandwich variance
estimator in addition to the model-based estimator.}

\item{debug}{Logical indicator to print diagnostic information}

\item{phi_init}{Optional. Vector of initial parameter values for estimation.}

\item{...}{additional arguments to [stats::nlm()].}
}
\value{
An object of class [joint_sem]
}
\description{
Estimate a joint model using SEM constraints
}
\examples{
data(joint_example)
fit <- joint_sem(
  joint_example, endpoints = c("Y1", "Y2", "Y3_cat"))
fit
estimate_effects(fit, risk_difference = "Y3_cat")
}
