% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{joint_sem_sl}
\alias{joint_sem_sl}
\title{Wrapper function for \code{joint_sem} for super learning.}
\usage{
joint_sem_sl(x, y, yname, treatment, categorical, phi_init = NULL, ...)
}
\arguments{
\item{x}{Matrix of secondary endpoints and treatment indicator}

\item{y}{Vector of primary endpoint measurements}

\item{yname}{Character name for primary endpoint}

\item{treatment}{Character name of the column name corresponding to binary
0-1 treatment indicator.}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}

\item{phi_init}{Optional. Vector of initial parameter values for estimation.}

\item{...}{additional arguments to \code{\link[stats:nlm]{stats::nlm()}}.}
}
\value{
An object of class \link{joint_sem}

An object of class {"joint_sem"} is a list containing at least the
following components:
\item{estimate}{The vector of estimated model parameters}
\item{ll}{The maximized log likelihood}
\item{dim_phi}{The number of model parameters}
\item{runtime}{The run time in seconds}
\item{endpoints}{The input vector of endpoints}
\item{categorical}{The vector of endpoints treated as categorical}
\item{V_model}{The model-based parameter variance estimate}
\item{nlm_code}{An integer indicating why the optimization process
terminated; see {"code"} from \code{\link[stats:nlm]{stats::nlm()}}}
\item{nlm_iterations}{The number of \code{\link[stats:nlm]{stats::nlm()}} iterations performed}
\item{call}{The matched call}
}
\description{
Wrapper function for \code{joint_sem} for super learning.
}
