% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner.R
\name{joint_vglm_sl}
\alias{joint_vglm_sl}
\title{Wrapper function for \code{VGAM::vglm} for Super Learning.}
\usage{
joint_vglm_sl(x, y, yname, treatment, ...)
}
\arguments{
\item{x}{Matrix of secondary endpoints and treatment indicator}

\item{y}{Vector of primary endpoint measurements}

\item{yname}{Character name for primary endpoint}

\item{categorical}{Character vector of endpoint names to be treated as
categorical (binary or ordinal) variables}
}
\description{
Wrapper function for \code{VGAM::vglm} for Super Learning.
}
